feat(llm): add full support of OpenAI API
